backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aioredis>=2.0.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "celery[redis]>=5.3.1",
    "daphne>=4.2.1",
    "fastapi>=0.122.0",
    "gunicorn>=24.1.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
    "ruff>=0.4.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # длину строки при желании пусть решает форматтер
    "F821",  # отключить "undefined name"
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
